# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Location Auto Maroc API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cars:
    get:
      operationId: listCars
      tags: [cars]
      summary: List cars (public, with filters)
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: brand
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: transmission
          in: query
          schema: { type: string }
        - name: fuelType
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: seats
          in: query
          schema: { type: integer }
        - name: available
          in: query
          schema: { type: boolean }
        - name: sortBy
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarListResponse"
    post:
      operationId: createCar
      tags: [cars]
      summary: Create a car (admin/agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
  /cars/{id}:
    get:
      operationId: getCar
      tags: [cars]
      summary: Get car details (public)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Car details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDetail"
    patch:
      operationId: updateCar
      tags: [cars]
      summary: Update a car (admin/agent)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
    delete:
      operationId: deleteCar
      tags: [cars]
      summary: Delete a car (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /cars/{id}/images:
    post:
      operationId: uploadCarImage
      tags: [cars]
      summary: Upload car image URL
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarImageInput"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarImage"
  /cars/{id}/availability:
    get:
      operationId: getCarAvailability
      tags: [cars]
      summary: Get car availability blocks
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Availability blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilityBlock"

  # ─── RENTAL REQUESTS ─────────────────────────────────────────────────────────
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts (public)
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get blog post by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}/edit:
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Car:
      type: object
      required: [id, brand, model, year, category, fuelType, transmission, seats, doors, dailyPrice, status]
      properties:
        id: { type: integer }
        brand: { type: string }
        model: { type: string }
        year: { type: integer }
        category: { type: string }
        fuelType: { type: string }
        transmission: { type: string }
        seats: { type: integer }
        doors: { type: integer }
        airConditioning: { type: boolean }
        dailyPrice: { type: number }
        weeklyPrice: { type: ["number", "null"] }
        monthlyPrice: { type: ["number", "null"] }
        depositAmount: { type: ["number", "null"] }
        mileageLimit: { type: ["integer", "null"] }
        city: { type: string }
        internalReference: { type: ["string", "null"] }
        licensePlate: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        seoSlug: { type: ["string", "null"] }
        insuranceIncluded: { type: boolean }
        requiredDocuments: { type: ["string", "null"] }
        status: { type: string }
        mainImageUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    CarDetail:
      type: object
      required: [id, brand, model, year, category, fuelType, transmission, seats, doors, dailyPrice, status]
      properties:
        id: { type: integer }
        brand: { type: string }
        model: { type: string }
        year: { type: integer }
        category: { type: string }
        fuelType: { type: string }
        transmission: { type: string }
        seats: { type: integer }
        doors: { type: integer }
        airConditioning: { type: boolean }
        dailyPrice: { type: number }
        weeklyPrice: { type: ["number", "null"] }
        monthlyPrice: { type: ["number", "null"] }
        depositAmount: { type: ["number", "null"] }
        mileageLimit: { type: ["integer", "null"] }
        city: { type: string }
        internalReference: { type: ["string", "null"] }
        licensePlate: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        seoSlug: { type: ["string", "null"] }
        insuranceIncluded: { type: boolean }
        requiredDocuments: { type: ["string", "null"] }
        status: { type: string }
        mainImageUrl: { type: ["string", "null"] }
        images:
          type: array
          items:
            $ref: "#/components/schemas/CarImage"
        createdAt: { type: string }
    CarListResponse:
      type: object
      required: [cars, total, page, limit]
      properties:
        cars:
          type: array
          items:
            $ref: "#/components/schemas/Car"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    CarInput:
      type: object
      required: [brand, model, year, category, fuelType, transmission, seats, doors, dailyPrice, city]
      properties:
        brand: { type: string }
        model: { type: string }
        year: { type: integer }
        category: { type: string }
        fuelType: { type: string }
        transmission: { type: string }
        seats: { type: integer }
        doors: { type: integer }
        airConditioning: { type: boolean }
        dailyPrice: { type: number }
        weeklyPrice: { type: number }
        monthlyPrice: { type: number }
        depositAmount: { type: number }
        mileageLimit: { type: integer }
        city: { type: string }
        internalReference: { type: string }
        licensePlate: { type: string }
        description: { type: string }
        seoTitle: { type: string }
        seoDescription: { type: string }
        seoSlug: { type: string }
        insuranceIncluded: { type: boolean }
        requiredDocuments: { type: string }
        status: { type: string }
        mainImageUrl: { type: string }
    CarUpdate:
      type: object
      properties:
        brand: { type: string }
        model: { type: string }
        year: { type: integer }
        category: { type: string }
        fuelType: { type: string }
        transmission: { type: string }
        seats: { type: integer }
        doors: { type: integer }
        airConditioning: { type: boolean }
        dailyPrice: { type: number }
        weeklyPrice: { type: number }
        monthlyPrice: { type: number }
        depositAmount: { type: number }
        mileageLimit: { type: integer }
        city: { type: string }
        internalReference: { type: string }
        licensePlate: { type: string }
        description: { type: string }
        seoTitle: { type: string }
        seoDescription: { type: string }
        seoSlug: { type: string }
        insuranceIncluded: { type: boolean }
        requiredDocuments: { type: string }
        status: { type: string }
        mainImageUrl: { type: string }
    CarImage:
      type: object
      required: [id, carId, url]
      properties:
        id: { type: integer }
        carId: { type: integer }
        url: { type: string }
        altText: { type: ["string", "null"] }
        isMain: { type: boolean }
        sortOrder: { type: integer }
    CarImageInput:
      type: object
      required: [url]
      properties:
        url: { type: string }
        altText: { type: string }
        isMain: { type: boolean }
        sortOrder: { type: integer }
    AvailabilityBlock:
      type: object
      required: [id, carId, startDate, endDate, type, status]
      properties:
        id: { type: integer }
        carId: { type: integer }
        rentalRequestId: { type: ["integer", "null"] }
        startDate: { type: string }
        endDate: { type: string }
        type: { type: string }
        status: { type: string }
        expiresAt: { type: ["string", "null"] }

    # ── Rental Requests ───────────────────────────────────────────────────────
...
    BlogPost:
      type: object
      required: [id, title, slug, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: ["string", "null"] }
        content: { type: ["string", "null"] }
        coverImage: { type: ["string", "null"] }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    BlogListResponse:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    BlogPostInput:
      type: object
      required: [title, slug, content]
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        coverImage: { type: string }
        seoTitle: { type: string }
        seoDescription: { type: string }
        status: { type: string }
    BlogPostUpdate:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        coverImage: { type: string }
        seoTitle: { type: string }
        seoDescription: { type: string }
        status: { type: string }

    # ── Notifications ─────────────────────────────────────────────────────────
...